+++ .angular/cache/18.2.9/CK1/vite/deps/@angular_common_http.js
@@ -37,9 +37,10 @@ import {
   withNoXsrfProtection,
   withRequestsMadeViaParent,
   withXsrfConfiguration
-} from "./chunk-WSCIPHTD.js";
-import "./chunk-RSIACHFR.js";
-import "./chunk-C6NMJ6KD.js";
+} from "./chunk-LK5GROIJ.js";
+import "./chunk-TO5JRZH6.js";
+import "./chunk-7ZEA7CND.js";
+import "./chunk-66JPAEYR.js";
 import "./chunk-EIB7IA3J.js";
 export {
   FetchBackend,
