+++ .angular/cache/18.2.9/CK1/vite/deps/@angular_material_button.js
@@ -4,8 +4,8 @@ import {
   MatRippleLoader,
   MatRippleModule,
   Platform
-} from "./chunk-CDNFK7WG.js";
-import "./chunk-RSIACHFR.js";
+} from "./chunk-AMJ7M4T2.js";
+import "./chunk-TO5JRZH6.js";
 import {
   ANIMATION_MODULE_TYPE,
   ChangeDetectionStrategy,
@@ -40,7 +40,8 @@ import {
   ɵɵinvalidFactory,
   ɵɵprojection,
   ɵɵprojectionDef
-} from "./chunk-C6NMJ6KD.js";
+} from "./chunk-7ZEA7CND.js";
+import "./chunk-66JPAEYR.js";
 import {
   __spreadProps,
   __spreadValues
