+++ .angular/cache/18.2.9/CK1/vite/deps/@angular_material_icon.js
@@ -1,15 +1,15 @@
 import {
   MatCommonModule
-} from "./chunk-CDNFK7WG.js";
+} from "./chunk-AMJ7M4T2.js";
 import {
   DomSanitizer
-} from "./chunk-AV3RCU74.js";
+} from "./chunk-UYWV6CS5.js";
 import {
   HttpClient
-} from "./chunk-WSCIPHTD.js";
+} from "./chunk-LK5GROIJ.js";
 import {
   DOCUMENT
-} from "./chunk-RSIACHFR.js";
+} from "./chunk-TO5JRZH6.js";
 import {
   Attribute,
   ChangeDetectionStrategy,
@@ -24,20 +24,10 @@ import {
   Optional,
   SecurityContext,
   SkipSelf,
-  Subscription,
   ViewEncapsulation$1,
   booleanAttribute,
-  catchError,
-  finalize,
-  forkJoin,
   inject,
-  map,
-  of,
   setClassMetadata,
-  share,
-  take,
-  tap,
-  throwError,
   ɵɵInputTransformsFeature,
   ɵɵStandaloneFeature,
   ɵɵattribute,
@@ -52,7 +42,19 @@ import {
   ɵɵinjectAttribute,
   ɵɵprojection,
   ɵɵprojectionDef
-} from "./chunk-C6NMJ6KD.js";
+} from "./chunk-7ZEA7CND.js";
+import {
+  Subscription,
+  catchError,
+  finalize,
+  forkJoin,
+  map,
+  of,
+  share,
+  take,
+  tap,
+  throwError
+} from "./chunk-66JPAEYR.js";
 import "./chunk-EIB7IA3J.js";
 
 // node_modules/@angular/material/fesm2022/icon.mjs
