+++ .angular/cache/18.2.9/CK1/vite/deps/@angular_material_toolbar.js
@@ -1,10 +1,10 @@
 import {
   MatCommonModule,
   Platform
-} from "./chunk-CDNFK7WG.js";
+} from "./chunk-AMJ7M4T2.js";
 import {
   DOCUMENT
-} from "./chunk-RSIACHFR.js";
+} from "./chunk-TO5JRZH6.js";
 import {
   ChangeDetectionStrategy,
   Component,
@@ -29,7 +29,8 @@ import {
   ɵɵprojection,
   ɵɵprojectionDef,
   ɵɵqueryRefresh
-} from "./chunk-C6NMJ6KD.js";
+} from "./chunk-7ZEA7CND.js";
+import "./chunk-66JPAEYR.js";
 import "./chunk-EIB7IA3J.js";
 
 // node_modules/@angular/material/fesm2022/toolbar.mjs
