+++ .angular/cache/18.2.9/CK1/vite/deps/@angular_platform-browser.js
@@ -35,12 +35,13 @@ import {
   withHttpTransferCacheOptions,
   withI18nSupport,
   withNoHttpTransferCache
-} from "./chunk-AV3RCU74.js";
-import "./chunk-WSCIPHTD.js";
+} from "./chunk-UYWV6CS5.js";
+import "./chunk-LK5GROIJ.js";
 import {
   getDOM
-} from "./chunk-RSIACHFR.js";
-import "./chunk-C6NMJ6KD.js";
+} from "./chunk-TO5JRZH6.js";
+import "./chunk-7ZEA7CND.js";
+import "./chunk-66JPAEYR.js";
 import "./chunk-EIB7IA3J.js";
 export {
   BrowserModule,
