+++ .angular/cache/18.2.9/CK1/vite/deps/_metadata.json
@@ -1,94 +1,103 @@
 {
-  "hash": "de37d772",
-  "configHash": "c5137282",
+  "hash": "e56e9644",
+  "configHash": "3c9be4d5",
   "lockfileHash": "db38567a",
-  "browserHash": "9a9763f1",
+  "browserHash": "af8233d8",
   "optimized": {
     "@angular/common": {
       "src": "../../../../../../node_modules/@angular/common/fesm2022/common.mjs",
       "file": "@angular_common.js",
-      "fileHash": "0f008ce1",
+      "fileHash": "d16e5add",
       "needsInterop": false
     },
     "@angular/common/http": {
       "src": "../../../../../../node_modules/@angular/common/fesm2022/http.mjs",
       "file": "@angular_common_http.js",
-      "fileHash": "303b87d4",
+      "fileHash": "8e686dad",
       "needsInterop": false
     },
     "@angular/core": {
       "src": "../../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "11e06a7f",
+      "fileHash": "fd5f6e1e",
       "needsInterop": false
     },
     "@angular/forms": {
       "src": "../../../../../../node_modules/@angular/forms/fesm2022/forms.mjs",
       "file": "@angular_forms.js",
-      "fileHash": "6de2d6ad",
+      "fileHash": "006cc639",
       "needsInterop": false
     },
     "@angular/material/button": {
       "src": "../../../../../../node_modules/@angular/material/fesm2022/button.mjs",
       "file": "@angular_material_button.js",
-      "fileHash": "676e8fbd",
+      "fileHash": "b3864a0d",
       "needsInterop": false
     },
     "@angular/material/icon": {
       "src": "../../../../../../node_modules/@angular/material/fesm2022/icon.mjs",
       "file": "@angular_material_icon.js",
-      "fileHash": "f6a5979a",
+      "fileHash": "e48768d5",
       "needsInterop": false
     },
     "@angular/material/toolbar": {
       "src": "../../../../../../node_modules/@angular/material/fesm2022/toolbar.mjs",
       "file": "@angular_material_toolbar.js",
-      "fileHash": "c3e48a1c",
+      "fileHash": "d7ee0cda",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "ece9f357",
+      "fileHash": "031bc4b0",
       "needsInterop": false
     },
     "@angular/platform-browser/animations/async": {
       "src": "../../../../../../node_modules/@angular/platform-browser/fesm2022/animations/async.mjs",
       "file": "@angular_platform-browser_animations_async.js",
-      "fileHash": "6cb4a7f3",
+      "fileHash": "8306c5ea",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "112d3248",
+      "fileHash": "183182ba",
       "needsInterop": false
     },
     "crypto-js": {
       "src": "../../../../../../node_modules/crypto-js/index.js",
       "file": "crypto-js.js",
-      "fileHash": "036101c1",
+      "fileHash": "65c728fb",
       "needsInterop": true
+    },
+    "rxjs": {
+      "src": "../../../../../../node_modules/rxjs/dist/esm5/index.js",
+      "file": "rxjs.js",
+      "fileHash": "4c77400f",
+      "needsInterop": false
     }
   },
   "chunks": {
-    "browser-HHIJJX5T": {
-      "file": "browser-HHIJJX5T.js"
+    "browser-SVTWWNLZ": {
+      "file": "browser-SVTWWNLZ.js"
     },
-    "chunk-CDNFK7WG": {
-      "file": "chunk-CDNFK7WG.js"
+    "chunk-AMJ7M4T2": {
+      "file": "chunk-AMJ7M4T2.js"
     },
-    "chunk-AV3RCU74": {
-      "file": "chunk-AV3RCU74.js"
+    "chunk-UYWV6CS5": {
+      "file": "chunk-UYWV6CS5.js"
     },
-    "chunk-WSCIPHTD": {
-      "file": "chunk-WSCIPHTD.js"
+    "chunk-LK5GROIJ": {
+      "file": "chunk-LK5GROIJ.js"
     },
-    "chunk-RSIACHFR": {
-      "file": "chunk-RSIACHFR.js"
+    "chunk-TO5JRZH6": {
+      "file": "chunk-TO5JRZH6.js"
     },
-    "chunk-C6NMJ6KD": {
-      "file": "chunk-C6NMJ6KD.js"
+    "chunk-7ZEA7CND": {
+      "file": "chunk-7ZEA7CND.js"
+    },
+    "chunk-66JPAEYR": {
+      "file": "chunk-66JPAEYR.js"
     },
     "chunk-EIB7IA3J": {
       "file": "chunk-EIB7IA3J.js"
