+++ src/app/services/auth.service.ts
@@ -1,5 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
+import { catchError, Observable, throwError } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -13,9 +14,20 @@ export class AuthService {
 
   constructor() { }
 
-  auth(email:string, password:string){
+  /*auth(email:string, password:string){
     let data = `username=${email}&password=${password}&grant_type=password`
     return this.http.post(this.url, data, {responseType: "text"})
+  }
+    */
+
+  auth(email: string, password: string): Observable<string> {
+    const data = `username=${email}&password=${password}&grant_type=password`;
+    return this.http.post(this.url, data, { responseType: 'text' }).pipe(
+      catchError((error) => {
+        console.error('Error en la solicitud:', error);
+        return throwError(() => error); // Propaga el error al componente
+      })
+    );
   }
 
 }
